nettelen

+++ terminal/lib/ui/screen/settings/settings_screen.dart
@@ -3,6 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:model_flight_logbook/domain/entities/terminal_endpoint.dart';
 import 'package:model_flight_logbook/injector.dart';
 import 'package:model_flight_logbook/l10n/generated/app_localizations.dart';
+import 'package:model_flight_logbook/ui/utils/toast.dart';
 import 'package:model_flight_logbook/ui/widgets/mfl_text_form_field.dart';
 import 'package:model_flight_logbook/ui/screen/settings/cubit/settings_cubit.dart';
 import 'package:model_flight_logbook/ui/screen/settings/cubit/settings_state.dart';
@@ -51,10 +52,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
             return Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Text(
-                  'Server',
-                  style: Theme.of(context).textTheme.headlineSmall,
-                ),
+                Text('Server', style: Theme.of(context).textTheme.headlineSmall),
                 const Divider(),
                 FormField(
                   builder: (field) {
@@ -63,11 +61,54 @@ class _SettingsScreenState extends State<SettingsScreen> {
                         return ListTile(
                           title: Text('${te.config.airportname} (${te.config.terminalname})'),
                           subtitle: Text(te.apiEndpoint),
-                          trailing: IconButton(
-                            onPressed: () {
-                              context.read<SettingsCubit>().deleteEndpointAndSave(te);
-                            },
-                            icon: const Icon(Icons.delete),
+                          trailing: PopupMenuButton(
+                            itemBuilder: (context) => [
+                              PopupMenuItem(
+                                onTap: () async {},
+                                child: ListTile(
+                                  title: const Text('Aktualisieren'),
+                                  leading: const Icon(Icons.refresh),
+                                  onTap: () async {
+                                    final updated = await context.read<SettingsCubit>().reloadEndpointAndSave(te);
+                                    if (context.mounted) {
+                                      Navigator.of(context).pop();
+                                      if (updated) {
+                                        Toast.success(context: context, message: 'Erfolgreich');
+                                      } else {
+                                        Toast.error(context: context, message: 'Fehlgeschlagen');
+                                      }
+                                    }
+                                  },
+                                ),
+                              ),
+                              PopupMenuItem(
+                                onTap: () async {},
+                                child: ListTile(
+                                  title: const Text('Löschen'),
+                                  leading: const Icon(Icons.delete),
+                                  onTap: () async {
+                                    final doDelete = await showAdaptiveDialog(
+                                      context: context,
+                                      builder: (context) {
+                                        return AlertDialog(
+                                          content: Text(
+                                            'Verbindung "${te.config.airportname} (${te.config.terminalname})" endgültig löschen?',
+                                            textAlign: TextAlign.center,
+                                          ),
+                                          actions: [
+                                            TextButton(onPressed: () => Navigator.of(context).pop(false), child: const Text('Abbrechen')),
+                                            TextButton(onPressed: () => Navigator.of(context).pop(true), child: const Text('Löschen')),
+                                          ],
+                                        );
+                                      },
+                                    );
+                                    if (doDelete && context.mounted) {
+                                      context.read<SettingsCubit>().deleteEndpointAndSave(te);
+                                    }
+                                  },
+                                ),
+                              ),
+                            ],
                           ),
                         );
                       }).toList(),
@@ -87,10 +128,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
                 const SizedBox(
                   height: 50,
                 ),
-                Text(
-                  'Sonstige Einstellungen',
-                  style: Theme.of(context).textTheme.titleLarge,
-                ),
+                Text('Sonstige Einstellungen', style: Theme.of(context).textTheme.headlineSmall),
                 const Divider(),
                 MflTextFormField(
                   controller: _aminPinController,
